requirements.txt
numpy==2.1.2
Pillow==11.0.0
PyAudio==0.2.14
pyvirtualcam==0.12.0
